got a start on adding stack index to emit functions
Still have problem with using the allocated stack.  Not sure it is
needed now.  The linkage from runtime to set up the stacks isn’t
working.  Not sure how critical this is, since can just use the same
stack as the runtime.  Seems much simpler.  Why go to the trouble of
allocating this other stack?  Probably needed over the longer term but
can work around for now.