dropped in closure implementation block comment
Still needs much clean-up.  Need to keep clear distinction between the
closure object and the label.   During code-generation the form
(closure label var …) generates code that allocate an object with the
code label embedded in it, and the code for that object leaving a
pointer to the object in %eax.