Documented redesigned IL
This IL is based on Dybvig 1995.  It recombines the closure and code
forms into a single closure form.  This will eliminate the need to
float closures to the top — which does not work, in general, because it
can separate variables from the context that determines where they are
allocated or referenced.  Instead we will expand closures inline with a
block of code that executes when the closure is called, and code to
build the closure object in the heap.  A second change is that variable
references are expanded to free, bound, and local forms.