fixed subtle bug in or
Subtle bug in or.  It was evaluating an argument twice.  Had to do with
how i transformed it to an if  (or X Y) => (if X X Y) is wrong if
evaluating X has side effects.  Used a let in the rewrite rule to make
sure argument on only evaluated once.  TRICKY!