got string->symbol body working in test cases 2.8
need to think through how I am going to associate the body with the
symbol.  how will code generation know that some funny symbol is going
to be bound to a closure?  should we rework all primitives?  define a
primitive to dereference the symbol like symbol-value.  Lots to think
through.