passing all regression tests now
replace the eliminate-shadowing transform with a simpler transform to
uniquify-variables.  Helped to design the transform with a simpler math
inspired notation; then hand compile that notation into scheme.  Found
a much simpler way to code the transform.  Took days of thinking, but
the code worked the first time!  Amazing how often that happens.